Moved moved table internals to private-components folder
