+++ frontend/src/lib/components/Table/index.ts
@@ -1,1 +1,2 @@
+export type { TableProps } from "./table";
 export { Table } from "./table";
